moved EGDS specified variables and functions to egds.go
